fix: feature data unwrapping for low and high pressure sodium

+++ app/src/main/java/com/gemini/energy/service/device/lighting/HighPressureSodium.kt
@@ -80,9 +80,9 @@ class HighPressureSodium(computable: Computable<*>, utilityRateGas: UtilityRate,
             val config = lightingConfig(ELightingType.HighPressureSodium)
             percentPowerReduced = config[ELightingIndex.PercentPowerReduced.value] as Double
 
-            peakHours = (featureData["Peak Hours"]!! as Int).toDouble()
-            partPeakHours = (featureData["Part Peak Hours"]!! as Int).toDouble()
-            offPeakHours = (featureData["Off Peak Hours"]!! as Int).toDouble()
+            peakHours = featureData["Peak Hours"]!! as Double
+            partPeakHours = featureData["Part Peak Hours"]!! as Double
+            offPeakHours = featureData["Off Peak Hours"]!! as Double
 
             alternateActualWatts = featureData["Alternate Actual Watts"]!! as Double
             alternateNumberOfFixtures = featureData["Alternate Number of Fixtures"]!! as Int
